Move codegen code to /codegen package
