--- src/local-state/codegen/index.ts
@@ -1,2 +1,0 @@
-export type { LocalStatePluginConfig } from "./config.js";
-export { plugin } from "./plugin.js";
